Revert "Develop.replace calendar"
