+++ frontend/package.json
@@ -14,13 +14,6 @@
     "@amcharts/amcharts5": "^5.10.7",
     "@emotion/react": "^11.14.0",
     "@emotion/styled": "^11.14.0",
-    "@fullcalendar/common": "^5.11.5",
-    "@fullcalendar/core": "^6.1.15",
-    "@fullcalendar/daygrid": "^6.1.15",
-    "@fullcalendar/interaction": "^6.1.15",
-    "@fullcalendar/list": "^6.1.15",
-    "@fullcalendar/react": "^6.1.15",
-    "@fullcalendar/timegrid": "^6.1.15",
     "@mui/icons-material": "^6.4.0",
     "@reduxjs/toolkit": "^2.3.0",
     "axios": "^1.7.7",
